Fix sing of `sizeof` in `tomurcuk::Bytes`

+++ public/tomurcuk/base/Bytes.hpp
@@ -45,9 +45,9 @@ namespace tomurcuk {
          */
         template<typename Element>
         static auto resetArray(Element *array, int64_t count) -> void {
-            assert(count <= INT64_MAX / sizeof(Element));
+            assert(count <= INT64_MAX / (int64_t)sizeof(Element));
 
-            resetBlock(array, count * sizeof(Element));
+            resetBlock(array, count * (int64_t)sizeof(Element));
         }
 
         /**
@@ -62,9 +62,9 @@ namespace tomurcuk {
          */
         template<typename Element>
         static auto resetSecretArray(Element *array, int64_t count) -> void {
-            assert(count <= INT64_MAX / sizeof(Element));
+            assert(count <= INT64_MAX / (int64_t)sizeof(Element));
 
-            resetSecretBlock(array, count * sizeof(Element));
+            resetSecretBlock(array, count * (int64_t)sizeof(Element));
         }
 
         /**
@@ -79,9 +79,9 @@ namespace tomurcuk {
          */
         template<typename Element>
         static auto copyArray(Element *destinationArray, Element *sourceArray, int64_t count) -> void {
-            assert(count <= INT64_MAX / sizeof(Element));
+            assert(count <= INT64_MAX / (int64_t)sizeof(Element));
 
-            copyBlock(destinationArray, sourceArray, count * sizeof(Element));
+            copyBlock(destinationArray, sourceArray, count * (int64_t)sizeof(Element));
         }
 
         /**
@@ -96,9 +96,9 @@ namespace tomurcuk {
          */
         template<typename Element>
         static auto copyAliasingArray(Element *destinationArray, Element *sourceArray, int64_t count) -> void {
-            assert(count <= INT64_MAX / sizeof(Element));
+            assert(count <= INT64_MAX / (int64_t)sizeof(Element));
 
-            copyAliasingBlock(destinationArray, sourceArray, count * sizeof(Element));
+            copyAliasingBlock(destinationArray, sourceArray, count * (int64_t)sizeof(Element));
         }
 
         /**
@@ -115,10 +115,10 @@ namespace tomurcuk {
          */
         template<typename Element>
         static auto testArrayExactness(Element *array0, int64_t count0, Element *array1, int64_t count1) -> bool {
-            assert(count0 <= INT64_MAX / sizeof(Element));
-            assert(count1 <= INT64_MAX / sizeof(Element));
+            assert(count0 <= INT64_MAX / (int64_t)sizeof(Element));
+            assert(count1 <= INT64_MAX / (int64_t)sizeof(Element));
 
-            return testBlockExactness(array0, count0 * sizeof(Element), array1, count1 * sizeof(Element));
+            return testBlockExactness(array0, count0 * (int64_t)sizeof(Element), array1, count1 * (int64_t)sizeof(Element));
         }
 
         /**
